@@ -1,8 +1,8 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.scss';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App title={'I am from props!'} />, document.getElementById('root'));
+createRoot(document.getElementById('root')).render(<App title={'I am from props!'} />);
 registerServiceWorker();
