@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import classes from './Car.module.scss';
 import withClass from '../hoc/withClass';
 import PropTypes from 'prop-types';
@@ -8,29 +8,29 @@ class Car extends React.Component {
   constructor(props) {
     super(props);
 
-    this.inputRef = React.createRef()
+    this.inputRef = React.createRef();
   }
 
   componentDidMount() {
-    if (this.props.index === 0) this.inputRef.current.focus()
+    if (this.props.index === 0) this.inputRef.current.focus();
   }
 
   render() {
 
-    const inputClasses = [classes.input]
+    const inputClasses = [classes.input];
 
     if (this.props.name !== '') {
-      inputClasses.push(classes.green)
+      inputClasses.push(classes.green);
     } else {
-      inputClasses.push(classes.red)
+      inputClasses.push(classes.red);
     }
 
     if (this.props.name.length > 4) {
-      inputClasses.push(classes.bold)
+      inputClasses.push(classes.bold);
     }
 
     return (
-      <Fragment>
+      <>
         <h3>Сar name: {this.props.name}</h3>
         <p>Year: <strong>{this.props.year}</strong></p>
         <input
@@ -41,7 +41,7 @@ class Car extends React.Component {
           className={inputClasses.join(' ')}
         />
         <button onClick={this.props.onDelete}>Delete</button>
-      </Fragment>
+      </>
     )
   }
 }
@@ -49,7 +49,7 @@ class Car extends React.Component {
 Car.propTypes = {
   name: PropTypes.string,
   year: PropTypes.number,
-  onChangeName: PropTypes.func
-}
+  onChangeName: PropTypes.func,
+};
 
-export default withClass(Car, classes.Car)
+export default withClass(Car, classes.Car);
