@@ -1,20 +1,20 @@
-import React, {Component} from 'react'
-import Auxiliary from '../hoc/Auxiliary'
-import Counter2 from '../Counter2/Counter2'
+import React, { Component } from 'react';
+import Auxiliary from '../hoc/Auxiliary';
+import Counter2 from '../Counter2/Counter2';
 
 export default class Counter extends Component {
   state = {
-    counter: 0
-  }
+    counter: 0,
+  };
 
   addCounter = () => {
 
     this.setState((prevState) => {
       return {
-        counter: prevState.counter + 1
+        counter: prevState.counter + 1,
       }
-    })
-  }
+    });
+  };
 
   render() {
     return (
@@ -22,8 +22,8 @@ export default class Counter extends Component {
         <h2>Counter {this.state.counter}</h2>
         <Counter2 />
         <button onClick={this.addCounter}>+</button>
-        <button onClick={() => this.setState({counter: this.state.counter - 1})}>-</button>
+        <button onClick={() => this.setState({ counter: this.state.counter - 1 })}>-</button>
       </Auxiliary>
     )
   }
-}
+};
