@@ -1,6 +1,4 @@
-// import React from 'react'
-
 // Like Fragment
-const Auxiliary = props => props.children
+const Auxiliary = props => props.children;
 
-export default Auxiliary
+export default Auxiliary;
